Improvement for Windows
First, awesome app!

I've wanted to use it on windows and ran in a problem and tried serving the app with waitress. The problem was with the ip 0.0.0.0, changing it to 127.0.0.1 fixed it. When reverting the app to default, I noticed that serving with waitress was way faster the via app.run. Thus the pull request.